Add flyway migration, Create class controller and service. Added log4j.properties

+++ src/main/java/ru/sladkkov/ChatSimbirSoft/domain/BotReactions.java
@@ -4,10 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Data
 @NoArgsConstructor
@@ -16,7 +13,7 @@ import javax.persistence.Id;
 public class BotReactions {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int reactionId;
+    private Long reactionId;
 
     private String outputMessage;
 }
